Add --script-cmd opt
Allows configuring a custom script to be executed for running the specs.

In this scenario, if this is specific to a particular TEST_REPO, then
this will only work with said repo and the user will have to configure
the cross repo run for that test in particular, and leave the others as
the default.