[CrossRepo::Repository] Add retry to ensure_clone
There have been a few cases where if many `ensure_clones` needed to be
called, then there is a chance for one of them to fail:

- https://app.travis-ci.com/github/ManageIQ/manageiq-cross_repo-tests/jobs/529191010#L325-L344
- https://app.travis-ci.com/github/ManageIQ/manageiq-cross_repo-tests/jobs/529191017#L326-L345

This, it is assumed, that there is a chance for network issues to cause
the download to fail, or be incomplete.

* * *

The simplest way to handle this is to just retry the download and see if
it works the second/third time around.