Unpack the hash for ruby 3 support
In prior rubies, this hash was unpacked as the only value in an array. Ruby 3 no
longer accepts kwargs or implicit hashes as the last argument sent to a method.

Since Runner#initialize accepts only kwargs, it's far easier to just change this
to unpack the hash into kwargs using double splat.
This was tested for for ruby 2.6, 2.7, and 3.0.